move Connections to mitmproxy.connection
