+++ mitmproxy/flow.py
@@ -2,11 +2,10 @@ import time
 import typing  # noqa
 import uuid
 
-from mitmproxy import controller
+from mitmproxy import controller, connection
 from mitmproxy import exceptions
 from mitmproxy import stateobject
 from mitmproxy import version
-from mitmproxy.proxy import context
 
 
 class Error(stateobject.StateObject):
@@ -55,18 +54,17 @@ class Error(stateobject.StateObject):
 
 
 class Flow(stateobject.StateObject):
-
     """
     A Flow is a collection of objects representing a single transaction.
     This class is usually subclassed for each protocol, e.g. HTTPFlow.
     """
 
     def __init__(
-            self,
-            type: str,
-            client_conn: context.Client,
-            server_conn: context.Server,
-            live: bool=None
+        self,
+        type: str,
+        client_conn: connection.Client,
+        server_conn: connection.Server,
+        live: bool = None
     ) -> None:
         self.type = type
         self.id = str(uuid.uuid4())
@@ -85,8 +83,8 @@ class Flow(stateobject.StateObject):
     _stateobject_attributes = dict(
         id=str,
         error=Error,
-        client_conn=context.Client,
-        server_conn=context.Server,
+        client_conn=connection.Client,
+        server_conn=connection.Server,
         type=str,
         intercepted=bool,
         is_replay=str,
