+++ test/mitmproxy/addons/test_block.py
@@ -1,7 +1,7 @@
 import pytest
 
+from mitmproxy import connection
 from mitmproxy.addons import block
-from mitmproxy.proxy import context
 from mitmproxy.test import taddons
 
 
@@ -27,7 +27,6 @@ from mitmproxy.test import taddons
     (True, False, True, ("::ffff:216.58.207.174",)),
     (True, False, True, ("2001:4860:4860::8888",)),
     (True, False, True, (r"2001:4860:4860::8888%scope",)),
-
 
     # block_private: loopback
     (False, True, False, ("127.0.0.1",)),
@@ -56,6 +55,6 @@ async def test_block_global(block_global, block_private, should_be_killed, addre
     ar = block.Block()
     with taddons.context(ar) as tctx:
         tctx.configure(ar, block_global=block_global, block_private=block_private)
-        client = context.Client(address, ("127.0.0.1", 8080), 1607699500)
+        client = connection.Client(address, ("127.0.0.1", 8080), 1607699500)
         ar.client_connected(client)
         assert bool(client.error) == should_be_killed
