+++ test/mitmproxy/net/http/test_headers.py
@@ -1,8 +1,10 @@
 import collections
+
 import pytest
 
 from mitmproxy.http import Headers
 from mitmproxy.net.http.headers import parse_content_type, assemble_content_type
+
 
 class TestHeaders:
     def _2host(self):
@@ -79,4 +81,5 @@ def test_assemble_content_type():
     p = assemble_content_type
     assert p("text", "html", {}) == "text/html"
     assert p("text", "html", {"charset": "utf8"}) == "text/html; charset=utf8"
-    assert p("text", "html", collections.OrderedDict([("charset", "utf8"), ("foo", "bar")])) == "text/html; charset=utf8; foo=bar"
+    assert p("text", "html",
+             collections.OrderedDict([("charset", "utf8"), ("foo", "bar")])) == "text/html; charset=utf8; foo=bar"
