+++ test/mitmproxy/proxy/layers/test_websocket.py
@@ -8,7 +8,7 @@ import wsproto.events
 from mitmproxy.http import HTTPFlow, Request, Response
 from mitmproxy.proxy.layers.http import HTTPMode
 from mitmproxy.proxy.commands import SendData, CloseConnection, Log
-from mitmproxy.proxy.context import ConnectionState
+from mitmproxy.connection import ConnectionState
 from mitmproxy.proxy.events import DataReceived, ConnectionClosed
 from mitmproxy.proxy.layers import http, websocket
 from mitmproxy.websocket import WebSocketFlow
